+++ apps/web/app/(admin)/members/_components/sidepanel/Form.tsx
@@ -1,8 +1,8 @@
-import { Controller } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { Controller, useForm } from "react-hook-form";
 import { Radio } from "@ui/index";
 import Input from "@ui/src/components/common/Input";
 import Button from "@ui/src/components/common/Button";
-import { REGEXP_PATTERNS } from "@repo/constants/regexp";
 import { MEMBER_FORM_MESSAGES } from "@repo/constants/messages";
 import {
   ROLE_LABELS,
@@ -10,8 +10,12 @@ import {
   type FormImageType,
   type MemberWithFileImage,
   type ImageFileType,
+  type SidePanelFormData,
 } from "@repo/types/src/membersType";
-import { useMembersForm } from "../../_hooks/useMembersForm";
+import { DEFAULT_VALUES } from "@repo/constants";
+import { useEffect } from "react";
+import { useMembersMutations } from "../../_hooks/useMembersMutations";
+import { memberFormSchema } from "../../_schemas/Form.schema";
 import ProfileImageUploader from "./ProfileImageUploader";
 import TeamDropdown from "./TeamDropdown";
 
@@ -28,9 +32,30 @@ export default function MemberForm({ selectedMember, onClose }: MemberFormProps)
     control,
     setValue,
     watch,
-    onSubmit: membersFormSubmit,
-    isPending,
-  } = useMembersForm({ selectedMember, onClose });
+    reset,
+  } = useForm({ defaultValues: DEFAULT_VALUES, resolver: zodResolver(memberFormSchema) });
+  const { updateMember, createMember, isPending } = useMembersMutations({
+    onSuccess: () => {
+      reset(DEFAULT_VALUES);
+      onClose();
+    },
+  });
+
+  const getRoleDisplay = (value: RoleOption): string => {
+    return ROLE_LABELS[value];
+  };
+
+  const getRoleValue = (displayText: string): RoleOption => {
+    const entry = Object.entries(ROLE_LABELS).find(([_, value]) => value === displayText);
+
+    return entry?.[0] as RoleOption;
+  };
+
+  const getCurrentImage = (): FormImageType => {
+    const profileImage = watch("profileImage");
+
+    return profileImage;
+  };
 
   const handleImageChange = (file: ImageFileType): void => {
     setValue("profileImage", file);
@@ -40,33 +65,49 @@ export default function MemberForm({ selectedMember, onClose }: MemberFormProps)
     return selectedMember ? MEMBER_FORM_MESSAGES.BUTTON.SUBMIT.UPDATE : MEMBER_FORM_MESSAGES.BUTTON.SUBMIT.ADD;
   };
 
-  const getRoleValue = (displayText: string): RoleOption => {
-    const entry = Object.entries(ROLE_LABELS).find(([_, value]) => value === displayText);
+  const createMemberFormData = (data: SidePanelFormData): FormData => {
+    const formData = new FormData();
+    formData.append("role", data.role);
+    formData.append("name", data.name);
+    formData.append("email", data.email);
 
-    return entry?.[0] as RoleOption;
+    data.teams.forEach((team) => {
+      formData.append("teams[]", team);
+    });
+
+    if (data.profileImage instanceof File) {
+      formData.append("profileImage", data.profileImage);
+    }
+
+    return formData;
   };
 
-  const getRoleDisplay = (value: RoleOption): string => {
-    return ROLE_LABELS[value];
+  const onSubmit = (data: SidePanelFormData): void => {
+    if (isPending) return;
+
+    const formData = createMemberFormData(data);
+    if (selectedMember) {
+      updateMember({ id: selectedMember._id, data: formData });
+    } else {
+      createMember(formData);
+    }
   };
 
-  const getCurrentImage = (): FormImageType => {
-    const profileImage = watch("profileImage");
-
-    return profileImage;
-  };
+  useEffect(() => {
+    if (selectedMember) {
+      reset({ ...selectedMember });
+    } else {
+      reset(DEFAULT_VALUES);
+    }
+  }, [reset, selectedMember]);
 
   return (
-    <form
-      className="flex h-full flex-col justify-between"
-      onSubmit={(...args) => void handleSubmit(membersFormSubmit)(...args)}
-    >
+    <form className="flex h-full flex-col justify-between" onSubmit={(...args) => void handleSubmit(onSubmit)(...args)}>
       <div>
         <div className="w-154 mb-24">
           <Controller
             name="role"
             control={control}
-            rules={{ required: MEMBER_FORM_MESSAGES.VALIDATION.ROLE.REQUIRED }}
             render={({ field: { value, onChange } }) => (
               <Radio.Group
                 value={getRoleDisplay(value)}
@@ -80,28 +121,11 @@ export default function MemberForm({ selectedMember, onClose }: MemberFormProps)
             )}
           />
         </div>
-        <Input
-          placeholder={MEMBER_FORM_MESSAGES.PLACEHOLDER.NAME}
-          error={errors.name}
-          {...register("name", {
-            required: MEMBER_FORM_MESSAGES.VALIDATION.NAME.REQUIRED,
-            minLength: {
-              value: 2,
-              message: MEMBER_FORM_MESSAGES.VALIDATION.NAME.MIN_LENGTH,
-            },
-          })}
-        />
-        <Input
-          placeholder={MEMBER_FORM_MESSAGES.PLACEHOLDER.EMAIL}
-          error={errors.email}
-          {...register("email", {
-            required: MEMBER_FORM_MESSAGES.VALIDATION.EMAIL.REQUIRED,
-            pattern: {
-              value: REGEXP_PATTERNS.EMAIL,
-              message: MEMBER_FORM_MESSAGES.VALIDATION.EMAIL.PATTERN,
-            },
-          })}
-        />
+
+        <Input placeholder={MEMBER_FORM_MESSAGES.PLACEHOLDER.NAME} error={errors.name} {...register("name")} />
+
+        <Input placeholder={MEMBER_FORM_MESSAGES.PLACEHOLDER.EMAIL} error={errors.email} {...register("email")} />
+
         <div className="mb-24">
           <Controller
             name="teams"
@@ -109,6 +133,7 @@ export default function MemberForm({ selectedMember, onClose }: MemberFormProps)
             render={({ field: { value, onChange } }) => <TeamDropdown value={value} onSelect={onChange} />}
           />
         </div>
+
         <ProfileImageUploader currentImage={getCurrentImage()} onImageChange={handleImageChange} />
       </div>
